Implement main server functionality with member management and event saving features
